% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cor.R, R/sim_cor_param.R
\name{sim_cor}
\alias{sim_cor}
\title{Create correlation of specified size}
\usage{
sim_cor(vector1 = NULL, vector2 = NULL, r, shuffles = 1000,
  N = NULL, M1 = NULL, SD1 = NULL, M2 = NULL, SD2 = NULL)

sim_cor(vector1 = NULL, vector2 = NULL, r, shuffles = 1000,
  N = NULL, M1 = NULL, SD1 = NULL, M2 = NULL, SD2 = NULL)
}
\arguments{
\item{vector1}{Optional: a vector for which a correlated vector should
be created.}

\item{vector2}{Optional: a vector that should be reshuffled in order to
create a correlation with vector1.}

\item{r}{Desired size of the correlation. Only takes values between -1
and 1.}

\item{shuffles}{Specifies how many times vector2 should be shuffled at
every step of the function. Defaults to 1000.}

\item{N}{Optional: Size of the two vectors to be created.}

\item{M1}{Optional: mean of the first vector to be drawn from a normal
distribution.}

\item{SD1}{Optional: standard deviation of the first vector to be
drawn from a normal distribution.}

\item{M2}{Optional: mean of the second vector to be drawn from a
normal distribution. Takes value of M1 if not provided.}

\item{SD2}{Optional: standard deviation of the second vector to be
drawn from a normal distribution. Takes value of M1 if not
provided.}

\item{N}{Size of the two vectors to be created.}

\item{dist}{A character string determining the distribution the first
vector should be sampled from. Can be set to 'normal' or 'unif'. See
details.}

\item{arglist}{A list of arguments provided to the distribution function
that is used to create the first vector. See details.}

\item{r}{Desired size of the correlation. Only takes values between -1
and 1.}

\item{shuffles}{Specifies how many times vector2 should be shuffled at
every step of the function. Defaults to 1000.}

\item{dist2}{A character string determining the distribution the second
vector should be sampled from. Can be set to 'normal' or 'unif'.
See details. If not provided, the same distribution as for the first
vector is used.}

\item{arglist2}{Optional: a list of arguments provided to the
distribution function that is used to create the second
vector. See details. If not provided, the same parameters
as for the first vector are used.}
}
\description{
Create correlation of specified size

Create correlation of specified size (parameter input)
}
\details{
Description yet to come.

Description yet to come.
}
\examples{

cor.7 <- sim_cor(N = 50, M1 = 10, SD1 = 1, M2 = 4, SD2 = 2, r = .7,
                 shuffles = 1000)


cor.7 <- sim_cor(N = 50, M1 = 10, SD1 = 1, M2 = 4, SD2 = 2, r = .7,
                 shuffles = 1000)

}
\author{
Juli Tkotz \email{juliane.tkotz@hhu.de}

Juli Tkotz \email{juliane.tkotz@hhu.de}
}
