% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpsons_paradox.R
\name{simpsons_paradox}
\alias{simpsons_paradox}
\title{Create a Simpson's Paradox}
\usage{
simpsons_paradox(r_tot, r_sub, ngroups = NULL, nsubgroups = 50,
  means_subgroups = NULL, sd_subgroups = 1, means_subgroups_y = NULL,
  sd_subgroups_y = NULL)
}
\arguments{
\item{r_tot}{Initial Pearson correlation of group medians. Will be
altered during the shuffling process within subgroups,
so consider it as the general direction of the overall
correlation in the data. See details.}

\item{r_sub}{Pearson correlation within the subgroups. Actual
correlation within subgroups will be very close to the
one provided in the \code{r_sub}.}

\item{ngroups}{Number of subgroups present in the data. Must not be
provided when a vector of subgroup means is used
instead (see \code{means_subgroups} and details).}

\item{nsubgroups}{Number of cases in each subgroup. Defaults to 50.}

\item{means_subgroups}{A vector providing the x-mean for each subgroup.
Must not be used when \code{ngroups} is used
instead (see details).}

\item{sd_subgroups}{Standard deviation for the x value of each subgroup.
Can be either a vector containing standard
deviations for each subgroup or an integer. In the
latter case, each subgroup will have the standard
deviation specified by the integer. Defaults to
sd = 1 for each group.}

\item{means_subgroups_y}{An optional vector providing the y-mean for each
subgroup. Can only be used when
\code{means_subgroups} is provided. Per default,
same means as in \code{means_subgroups} are used.
Note that the order of y-means will be shuffled
in order to create a correlation.}

\item{sd_subgroups_y}{Standard deviation for the y value of each subgroup.
Can be either a vector containing standard
deviations for each subgroup or an integer. In the
latter case, each subgroup will have the standard
deviation specified by the integer. Per default
identical to \code{sd_subgroups}.}
}
\description{
Create a Simpson's Paradox
}
\details{
Creates a Simpson's Paradox by creating a correlation between the means
of two variables x and y of a number of subgroups. For this step, the
function \code{\link{sim_cor_vec}} is used.

Then creates a correlation within subgroups using
\code{\link{sim_cor_param}} based on the means created in the first step.
If the two correlations specified for each step (\code{r_tot} and
\code{r_sub}, respectively) have opposite directions, a Simpson's Paradox
is created.

One of either \code{ngroups} or \code{mean_subgroups} needs to be provided.
That is, if the number of subgroups is specified via \code{ngroups}, the
x-means for each group will be \code{1:ngroups}. If the x-means for each
subgroup are specified via the vector \code{means_subgroups}, the number
of subgroups equals the number of means provided in the
\code{means_subgroups}. You can specify the y-means of each subgroup via
\code{means_subgroups_y}, but since the y-means will be shuffled in oder
to create a correlation across the whole data, this rather sets a range
of values for the y-means instead of specific y-means for each group.

Based on linear correlations (pearson).

Returns a \code{data.frame} that holds the x and y coordinates of the
data, a \code{group} column containing the subgroup each case belongs
to and the correlation within each subgroup.

Depending on \code{r_tot}, \code{r_sub} and standard deviations, the
original correlation provided via \code{r_tot} might be substantially
reduced. Toy around with it and see which parameters yield the 'best'
Simpson's Paradox.
}
\examples{

simpson <- simpsons_paradox(r_tot = .9, r_sub = -.6, ngroups = 4, nsubgroups = 30)

}
\author{
Juli Tkotz \email{juliane.tkotz@hhu.de}
}
