% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cor.R
\name{sim_cor}
\alias{sim_cor}
\title{Create correlation of specified size}
\usage{
sim_cor(vector1 = NULL, vector2 = NULL, r, shuffles = 1000,
  N = NULL, MW1 = NULL, SD1 = NULL, MW2 = NULL, SD2 = NULL)
}
\arguments{
\item{vector1}{Optional: a vector for which a correlated vector should
be created.}

\item{vector2}{Optional: a vector that should be reshuffled in order to
create a correlation with vector1.}

\item{r}{Desired size of the correlation. Only takes values between -1
and 1.}

\item{shuffles}{Specifies how many times vector2 should be shuffled at
every step of the function. Defaults to 1000.}

\item{N}{Optional: Size of the two vectors to be created.}

\item{MW1}{Optional: mean of the first vector to be drawn from a normal
distribution.}

\item{SD1}{Optional: standard deviation of the first vector to be
drawn from a normal distribution.}

\item{MW2}{Optional: mean of the second vector to be drawn from a
normal distribution. Takes value of MW1 if not provided.}

\item{SD2}{Optional: standard deviation of the second vector to be
drawn from a normal distribution. Takes value of MW1 if not
provided.}
}
\description{
Create correlation of specified size
}
\details{
Description yet to come.
}
\examples{

cor.7 <- sim_cor(N = 50, MW1 = 10, SD1 = 1, MW2 = 4, SD2 = 2, r = .7,
                 shuffles = 1000)

}
\author{
Juli Tkotz \email{juliane.tkotz@hhu.de}
}
